+++ src/Entity/ClubDependent/MemberSeason.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Entity;
+namespace App\Entity\ClubDependent;
 
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\GetCollection;
 use ApiPlatform\Metadata\Link;
 use App\Entity\Abstract\UuidEntity;
-use App\Entity\ClubDependent\Member;
+use App\Entity\AgeCategory;
+use App\Entity\Season;
 use App\Repository\MemberSeasonRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Attribute\Groups;
