+++ src/Entity/User.php
@@ -20,7 +20,7 @@ use App\Entity\Abstract\UuidEntity;
 use App\Enum\UserRole;
 use App\Filter\MultipleFilter;
 use App\Repository\UserRepository;
-use App\State\MemberProcessor;
+use App\State\UserMemberProcessor;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
@@ -129,7 +129,7 @@ use Symfony\Component\Validator\Constraints as Assert;
   'groups' => ['user', 'user-read'],
 ], denormalizationContext: [
   'groups' => ['user', 'user-write'],
-], processor: MemberProcessor::class,)]
+], processor: UserMemberProcessor::class,)]
 #[ApiFilter(SearchFilter::class, properties: ['role' => 'exact'])]
 #[ApiFilter(OrderFilter::class, properties: ['lastname' => 'ASC', 'firstname' => 'ASC'])]
 #[ApiFilter(MultipleFilter::class, properties: ['firstname', 'lastname'])]
