+++ src/Repository/ActivityRepository.php
@@ -3,8 +3,8 @@
 namespace App\Repository;
 
 use App\Entity\ClubDependent\Activity;
+use App\Entity\ClubDependent\MemberPresence;
 use App\Entity\ExternalPresence;
-use App\Entity\MemberPresence;
 use App\Repository\Trait\UuidEntityRepositoryTrait;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
