+++ tests/Factory/MemberPresenceFactory.php
@@ -2,7 +2,7 @@
 
 namespace App\Tests\Factory;
 
-use App\Entity\MemberPresence;
+use App\Entity\ClubDependent\MemberPresence;
 use App\Repository\MemberPresenceRepository;
 use App\Tests\Story\ActivityStory;
 use DateTimeImmutable;
@@ -31,22 +31,22 @@ use Zenstruck\Foundry\Persistence\ProxyRepositoryDecorator;
  * @method FactoryCollection<MemberPresence|Proxy> sequence(iterable|callable $sequence)
  * @method static ProxyRepositoryDecorator<MemberPresence, MemberPresenceRepository> repository()
  *
- * @phpstan-method \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> create(array|callable $attributes = [])
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> createOne(array $attributes = [])
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> find(object|array|mixed $criteria)
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> findOrCreate(array $attributes)
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> first(string $sortedField = 'id')
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> last(string $sortedField = 'id')
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> random(array $attributes = [])
- * @phpstan-method static \App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence> randomOrCreate(array $attributes = [])
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> all()
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> createMany(int $number, array|callable $attributes = [])
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> createSequence(iterable|callable $sequence)
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> findBy(array $attributes)
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> randomRange(int $min, int $max, array $attributes = [])
- * @phpstan-method static list<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> randomSet(int $number, array $attributes = [])
- * @phpstan-method \Zenstruck\Foundry\FactoryCollection<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> many(int $min, int|null $max = null)
- * @phpstan-method \Zenstruck\Foundry\FactoryCollection<\App\Entity\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\MemberPresence>> sequence(iterable|callable $sequence)
+ * @phpstan-method \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> create(array|callable $attributes = [])
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> createOne(array $attributes = [])
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> find(object|array|mixed $criteria)
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> findOrCreate(array $attributes)
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> first(string $sortedField = 'id')
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> last(string $sortedField = 'id')
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> random(array $attributes = [])
+ * @phpstan-method static \App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence> randomOrCreate(array $attributes = [])
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> all()
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> createMany(int $number, array|callable $attributes = [])
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> createSequence(iterable|callable $sequence)
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> findBy(array $attributes)
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> randomRange(int $min, int $max, array $attributes = [])
+ * @phpstan-method static list<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> randomSet(int $number, array $attributes = [])
+ * @phpstan-method \Zenstruck\Foundry\FactoryCollection<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> many(int $min, int|null $max = null)
+ * @phpstan-method \Zenstruck\Foundry\FactoryCollection<\App\Entity\ClubDependent\MemberPresence&\Zenstruck\Foundry\Persistence\Proxy<\App\Entity\ClubDependent\MemberPresence>> sequence(iterable|callable $sequence)
  * @extends PersistentProxyObjectFactory<MemberPresence>
  */
 final class MemberPresenceFactory extends PersistentProxyObjectFactory {
